pyproject.toml
[project]
name = "crt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "moderngl>=5.12.0",
    "numpy>=2.3.4",
    "pygame-ce>=2.5.5",
]
